refactor: centralize Base definition
